services/ingestion-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
firebase-admin